Refactor file order and delete dir Controller
